requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langgraph>=0.0.26

# OpenAI
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Data Processing
sentence-transformers>=2.2.0

# Jupyter (for notebook support)
ipython>=8.0.0
jupyter>=1.0.0

# Document Processing (if using LlamaParse from the notebook)
llama-parse>=0.4.0

# Additional common dependencies that might be needed
numpy>=1.21.0
pandas>=1.3.0

# Type hints support
typing-extensions>=4.0.0

# HTTP requests (commonly needed by AI libraries)
requests>=2.25.0

# Async support (used by many AI libraries)
aiohttp>=3.8.0